requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-frontmatter>=1.1.0
python-slugify>=8.0.0
aiosqlite>=0.20.0
Jinja2>=3.1.0
Markdown>=3.7
# Cloud deployment
asyncpg>=0.30.0
httpx>=0.27.0
